projects/Marketingtool/backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# PPT Generation
python-pptx==0.6.23

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
